Modified quite a bit in the entity, controller, service, TO and converter directories.
Go through all or most of them and test their functionality if unsure whether it works or not.

Just tested the gameUpdate Put Request and it kind of work, but it should maybe be possible to change the game genre as well.

Also, go through all services and controllers and modify so all get requests return TOs not the original entity, and the post and put requests accept original entities and return TOs if they do so.
This needs to be discussed whether they should return or not ( technically not but we could leave it as is ).